Deal amount should be decimal not int

+++ src/Deal/Interfaces/IDealHubSpotEntity.cs
@@ -10,7 +10,7 @@ namespace Skarp.HubSpotClient.Deal.Interfaces
         string Pipeline { get; set; }
         string OwnerId { get; set; }
         string CloseDate { get; set; }
-        int Amount { get; set; }
+        decimal Amount { get; set; }
         string DealType { get; set; }
         string RouteBasePath { get; }
     }
